restructure lightgbm module
