+++ tests/lightgbm/test_export_tree_data.py
@@ -1,4 +1,4 @@
-from tabular_trees.lightgbm import LightGBMTabularTrees
+from tabular_trees.lightgbm.lightgbm_tabular_trees import LightGBMTabularTrees
 from tabular_trees.trees import export_tree_data
